version 1.1

+++ app/Http/Controllers/Admin/Panel/AdminsController.php
@@ -3,41 +3,28 @@
 namespace App\Http\Controllers\Admin\Panel;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\UpdateUserRequest;
-use App\Http\Requests\Admin\UserRequest;
 use App\Models\User;
 use App\Traits\AuthTrait;
 use App\Traits\GeneralTrait;
+use App\Traits\Messages\PanelMessagesTrait;
+use App\Traits\Rules\PanelRulesTrait;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
+use function Laravel\Prompts\select;
 
 class AdminsController extends Controller
 {
     use GeneralTrait;
     use AuthTrait;
+    use PanelRulesTrait;
+    use PanelMessagesTrait;
 
     public function index()
     {
         return $this->ifAdmin('admin.panel.admins.index', [
-                    'admins' => DB::table('users as admins')
-                    ->join('sectors', 'admins.sector_id', '=', 'sectors.id')
-                    ->join('lines', 'admins.line_id', '=', 'lines.id')
-                    ->join('titles', 'admins.title_id', '=', 'titles.id')
-                    ->where('role', 1)
-                    ->select(
-                        'admins.id',
-                        'admins.first_name',
-                        'admins.middle_name',
-                        'admins.last_name',
-                        'admins.user_name',
-                        'admins.email',
-                        'admins.phone_number',
-                        'admins.profile_image',
-                        'admins.activated',
-                        'admins.created_at',
-                        'sectors.name as sector_name',
-                        'lines.name as line_name',
-                        'titles.name as title_name',
-                    )
+            'admins' => DB::table('users')->where('role', 1)
+            ->select('id', 'first_name', 'email', 'profile_image', 'activated', 'created_at')
         ]);
     }
 
@@ -46,48 +33,59 @@ class AdminsController extends Controller
      */
     public function create()
     {
-        return $this->ifAdmin('admin.panel.admins.create', [
-                'sectors' => DB::table('sectors')->select(['id', 'name'])->get(),
-                'lines' => DB::table('lines')->select(['id', 'name'])->get(),
-                'titles' => DB::table('titles')->select(['id', 'name'])->get(),
-            ]);
+        if (auth()->id() == 1) {
+            return $this->ifAdmin('admin.panel.admins.create');
+        }
+        return back();
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(UserRequest $request)
+    public function store(Request $request)
     {
-        $admin = new User();
-        $data = $request->all();
+        if (auth()->id() == 1) {
+            try {
+                $admin = new User();
+                $data = $request->all();
 
-        $firstName = strtolower($data['first_name']);
-        $middleName = strtolower($data['middle_name']);
-        $username = $firstName . $middleName . rand(1, 9);
+                $validator = Validator::make($data, $this->adminRules(), $this->adminMessages());
 
-        $i = 0;
-        while (User::whereuser_name($username)->exists()) {
-            $i++;
-            $username .= $i;
+                if ($validator->fails()) {
+                    return $this->backWithMessage('error', $validator->errors()->first());
+                }
+
+                $name = strtolower($data['name']);
+                $code = $name . rand(1, 9);
+
+                $i = 0;
+                while (User::whereuser_name($code)->exists() || User::wherecrm_code($code)->exists() || User::wherephone_number($code)->exists()) {
+                    $i++;
+                    $code .= $i;
+                }
+
+                $admin->first_name = $data['name'];
+                $admin->middle_name = $data['name'];
+                $admin->last_name =  $data['name'];
+                $admin->user_name = $code;
+                $admin->crm_code = $code;
+                $admin->email = $data['email'];
+                $admin->phone_number = $code;
+                $admin->password = bcrypt($data['password']);
+                $admin->title_id = 0;
+                $admin->lines = 0;
+                $admin->sectors = 0;
+                $admin->role = 1;
+                $admin->activated = 1;
+
+                $admin->save();
+
+                return $this->backWithMessage('success', 'Admin added successfully');
+            } catch (\Exception $e) {
+                return $this->backWithMessage('error', 'Something went error, please try again later');
+            }
         }
-
-        $admin->first_name = $data['first_name'];
-        $admin->middle_name = $data['middle_name'];
-        $admin->last_name =  $data['last_name'];
-        $admin->user_name = $username;
-        $admin->crm_code = $data['crm_code'];
-        $admin->email = $data['email'];
-        $admin->phone_number = $data['phone_number'];
-        $admin->password = bcrypt($data['password']);
-        $admin->title_id = $data['title'];
-        $admin->line_id = $data['line'];
-        $admin->sector_id = $data['sector'];
-        $admin->role = 1;
-        $admin->activated = 1;
-
-        $admin->save();
-
-        return $this->backWithMessage('uploadedSuccessfully', 'Admin Added Successfully');
+        return back();
     }
 
     /**
@@ -95,37 +93,60 @@ class AdminsController extends Controller
      */
     public function edit(string $id)
     {
-        return $this->ifAdmin('admin.panel.admins.edit', [
-                'selected_admin' => DB::table('users')->where('id', '=', $id)->first(),
-                'sectors' => DB::table('sectors')->select(['id', 'name'])->get(),
-                'lines' => DB::table('lines')->select(['id', 'name'])->get(),
-                'titles' => DB::table('titles')->select(['id', 'name'])->get(),
+        if (auth()->id() == 1 || auth()->id() == $id) {
+            return $this->ifAdmin('admin.panel.admins.edit', [
+                'selected_admin' => DB::table('users')->where('id', '=', $id)->
+                select('id', 'first_name', 'email')->first(),
             ]);
+        }
+        return back();
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateUserRequest $request, string $id)
+    public function update(Request $request, string $id)
     {
-        DB::table('users')
-            ->where('id', '=', $id)
-            ->update([
-                'first_name' => $request->first_name,
-                'middle_name' => $request->middle_name,
-                'last_name' => $request->last_name,
-                'user_name' => $request->user_name,
-                'crm_code' => $request->crm_code,
-                'email' => $request->email,
-                'phone_number' => $request->phone_number,
-                'title_id' => $request->title,
-                'line_id' => $request->line,
-                'sector_id' => $request->sector,
-                'role'=> 1,
-                'activated' => 1,
-            ]);
+        if (auth()->id() == 1 || auth()->id() == $id) {
+            try {
+                $validator = Validator::make($request->all(), $this->adminUpdateRules($id), $this->adminUpdateMessages());
 
-        return $this->backWithMessage('savedSuccessfully', 'Admin Details saved successfully');
+                if ($validator->fails()) {
+                    return $this->backWithMessage('error', $validator->errors()->first());
+                }
+
+                $name = strtolower($request->name);
+                $code = $name . rand(1, 9);
+
+                $i = 0;
+                while (User::whereuser_name($code)->exists() || User::wherecrm_code($code)->exists() || User::wherephone_number($code)->exists()) {
+                    $i++;
+                    $code .= $i;
+                }
+
+                DB::table('users')
+                    ->where('id', '=', $id)
+                    ->update([
+                        'first_name' => $request->name,
+                        'middle_name' => $request->name,
+                        'last_name' => $request->name,
+                        'user_name' => $code,
+                        'crm_code' => $code,
+                        'email' => $request->email,
+                        'phone_number' => $code,
+                        'title_id' => 0,
+                        'lines' => 0,
+                        'sectors' => 0,
+                        'role'=> 1,
+                        'activated' => 1,
+                    ]);
+
+                return $this->backWithMessage('success', 'Admin details saved successfully');
+            } catch (\Exception $e) {
+                return $this->backWithMessage('error', 'Something went error, please try again later');
+            }
+        }
+        return back();
     }
 
     /**
@@ -133,7 +154,10 @@ class AdminsController extends Controller
      */
     public function destroy(string $id)
     {
-        $this->deleteFromDB('users', $id, 'storage/images/profile_images/', 'profile_image');
-        return $this->backWithMessage('deletedSuccessfully', 'Admin has been deleted');
+        if (auth()->id() == 1) {
+            $this->deleteFromDB('users', $id, 'storage/images/profile_images/', 'profile_image');
+            return $this->backWithMessage('success', 'Admin has been deleted');
+        }
+        return back();
     }
 }
