+++ app/Http/Controllers/Admin/Panel/TopicsController.php
@@ -3,17 +3,22 @@
 namespace App\Http\Controllers\Admin\Panel;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\TopicRequest;
 use App\Models\Topic;
 use App\Models\TopicNotification;
 use App\Traits\AuthTrait;
 use App\Traits\GeneralTrait;
+use App\Traits\Messages\PanelMessagesTrait;
+use App\Traits\Rules\PanelRulesTrait;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class TopicsController extends Controller
 {
     use GeneralTrait;
     use AuthTrait;
+    use PanelRulesTrait;
+    use PanelMessagesTrait;
 
     /**
      * Display a listing of the resource.
@@ -42,31 +47,41 @@ class TopicsController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(TopicRequest $request)
+    public function store(Request $request)
     {
-        $topic_title = str_replace(' ', '', $request->title);
-        $topic_image = $topic_title . time() . '.' . $request->image->extension();
+        try {
+            $validator = Validator::make($request->all(), $this->topicRules(), $this->topicMessages());
 
-        $topic = new Topic();
+            if ($validator->fails()) {
+                return $this->backWithMessage('error', $validator->errors()->first());
+            }
 
-        $topic->title = $request->title;
-        $topic->description = $request->description;
-        $topic->image = $topic_image;
-        $topic->user_id = auth()->user()->id;
-        $topic->status = 1;
+            $topic_title = str_replace(' ', '', $request->title);
+            $topic_image = $topic_title . time() . '.' . $request->image->extension();
 
-        $topic->save();
+            $topic = new Topic();
 
-        $request->image->storeAs('public/images/topics', $topic_image);
+            $topic->title = $request->title;
+            $topic->description = $request->description;
+            $topic->image = $topic_image;
+            $topic->user_id = auth()->user()->id;
+            $topic->status = 1;
 
-        $notification = new TopicNotification;
+            $topic->save();
 
-        $notification->text = auth()->user()->first_name . ' ' . auth()->user()->middle_name . ' added a new topic - ' . $request->title;
-        $notification->topic_id = DB::table('topics')->latest('id')->first()->id;
+            $request->image->storeAs('public/images/topics', $topic_image);
 
-        $notification->save();
+            $notification = new TopicNotification;
 
-        return $this->backWithMessage('uploadedSuccessfully', 'Topic Shared Successfully');
+            $notification->text = auth()->user()->first_name . ' ' . auth()->user()->middle_name . ' added a new topic - ' . $request->title;
+            $notification->topic_id = DB::table('topics')->latest('id')->first()->id;
+
+            $notification->save();
+
+            return $this->backWithMessage('success', 'Topic Shared Successfully');
+        } catch (\Exception $e) {
+            return $this->backWithMessage('error', 'Something went error, please try again later');
+        }
     }
 
     /**
@@ -83,39 +98,49 @@ class TopicsController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(TopicRequest $request, string $id)
+    public function update(Request $request, string $id)
     {
-        $old_topic_name = DB::table('topics')->select('title')->where('id', $id)->first();
+        try {
+            $validator = Validator::make($request->all(), $this->topicRules(), $this->topicMessages());
 
-        $storedImage = DB::table('topics')->select('image')->where('id', $id)->first();
-        if (file_exists(asset('storage/images/topics/' . $storedImage->image)))
-            unlink(asset('storage/images/topics/' . $storedImage->image));
+            if ($validator->fails()) {
+                return $this->backWithMessage('error', $validator->errors()->first());
+            }
 
-        $topic_title = str_replace(' ', '', $request->title);
-        $topic_image = $topic_title . time() . '.' . $request->image->extension();
+            $old_topic_name = DB::table('topics')->select('title')->where('id', $id)->first();
 
-        DB::table('topics')
-            ->where('id', '=', $id)
-            ->update([
-                'title' => $request->title,
-                'description' => $request->description,
-                'image' => $topic_image,
-            ]);
+            $storedImage = DB::table('topics')->select('image')->where('id', $id)->first();
+            if (file_exists(asset('storage/images/topics/' . $storedImage->image)))
+                unlink(asset('storage/images/topics/' . $storedImage->image));
 
-        $request->image->storeAs('public/images/topics', $topic_image);
+            $topic_title = str_replace(' ', '', $request->title);
+            $topic_image = $topic_title . time() . '.' . $request->image->extension();
 
-        $old_topic_name->title != $request->title ? $title_updated = 1 : $title_updated = 0;
+            DB::table('topics')
+                ->where('id', '=', $id)
+                ->update([
+                    'title' => $request->title,
+                    'description' => $request->description,
+                    'image' => $topic_image,
+                ]);
 
-        $notification = new TopicNotification;
+            $request->image->storeAs('public/images/topics', $topic_image);
 
-        $notification->text = $title_updated == 1 ?
-            auth()->user()->first_name . ' ' . auth()->user()->middle_name . ' updated ' . $old_topic_name->title .' topic to ' . $request->title :
-            auth()->user()->first_name . ' ' . auth()->user()->middle_name . ' updated ' . $old_topic_name->title .' topic';
-        $notification->topic_id = $id;
+            $old_topic_name->title != $request->title ? $title_updated = 1 : $title_updated = 0;
 
-        $notification->save();
+            $notification = new TopicNotification;
 
-        return $this->backWithMessage('savedSuccessfully', 'Topic saved successfully');
+            $notification->text = $title_updated == 1 ?
+                auth()->user()->first_name . ' ' . auth()->user()->middle_name . ' updated ' . $old_topic_name->title .' topic to ' . $request->title :
+                auth()->user()->first_name . ' ' . auth()->user()->middle_name . ' updated ' . $old_topic_name->title .' topic';
+            $notification->topic_id = $id;
+
+            $notification->save();
+
+            return $this->backWithMessage('success', 'Topic saved successfully');
+        } catch (\Exception $e) {
+            return $this->backWithMessage('error', 'Something went error, please try again later');
+        }
     }
 
     /**
@@ -125,6 +150,6 @@ class TopicsController extends Controller
     {
         $this->deleteFromDB('topics', $id, 'storage/images/topics/', 'image');
         DB::table('comments')->where('topic_id', $id)->delete();
-        return $this->backWithMessage('deletedSuccessfully', 'Topic has been deleted');
+        return $this->backWithMessage('success', 'Topic has been deleted');
     }
 }
